@@ -1,13 +1,9 @@
 import './globals.css';
-import { Inter } from 'next/font/google';
 import type { Metadata } from 'next';
 
-
-const inter = Inter({ subsets: ['latin'] });
-
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Excel Processor',
+  description: 'Process your Excel files with AI',
 };
 
 export default function RootLayout({
@@ -16,8 +12,12 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="en" className={inter.className}>
-      <body>{children}</body>
+    <html lang="en" className="sans">
+      <body className="min-h-screen bg-neutralColor">
+        <main className="flex flex-col">
+          {children}
+        </main>
+      </body>
     </html>
   );
 }
